Pool (#93)
* Removing unsafe from int e uint conversion (#26)

* Revert "Removing unsafe from int e uint conversion (#26)" (#27)

This reverts commit 018fe9fd95039b6fcaf26fc0e0ae64be91aa115e.

* IBinaryDeserializer

* Remove auto-warm

* Binary serializer changes

* IBinaryConverter

* DeserializeInside => Deserialize

- Remove async test

* Allow compression

* Optimize ProtocolV2

* Random in TestBase class

- Clean code of ConsoleReader

* RPC server

- Update packages
- Undone register some IRpcRequest

* Null check

* IRpcRequest => IRpcProcessRequest

* Clean code

* Clean code

* Clean code

* Default Values

* Clean code

* Clean code

* Easy style definition

* Console improves

* Reorder commands, improve help, allow enum in parameters

* Easy autocomplete "wallet" "rpc" and "network" commands

* fix

* Allow to write in multiple formats

* Header/quit/bLock

* TODOs

* TODO

* Asset, Contract a priority

* Sort usings

* Build contract

* Renames

* Fix

* Records

* change name

* Optimize event

* Fix Wallet close

* Improve parameter checks

* Clean code

* Pool

* Travis, try again please :P

*  Base58 UT and coments (#92) (#14)

* Removing unsafe from int e uint conversion (#26)

* Revert "Removing unsafe from int e uint conversion (#26)" (#27)

This reverts commit 018fe9fd95039b6fcaf26fc0e0ae64be91aa115e.

* ACL Unit Tests and some improvement

* Base58 UT and coments

* Update UtBase58.cs

* Update UtBase58.cs

* StampPool (with date)

* Null BC

* state

* Fix Travis StackOverflow

* Try again travis

+++ src/NeoSharp.Application/Client/ConsoleWriter.cs
@@ -80,7 +80,7 @@ namespace NeoSharp.Application.Client
         /// </summary>
         /// <param name="maxValue">Maximum value</param>
         /// <returns>Return Console percent writer</returns>
-        public ConsolePercentWriter CreatePercent(int maxValue = 100)
+        public ConsolePercentWriter CreatePercent(long maxValue = 100)
         {
             return new ConsolePercentWriter(this, 0, maxValue);
         }
